requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Werkzeug==2.3.7
alembic==1.12.1
pyodbc==5.0.1
python-dotenv==1.0.0 